% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paws.pi_package.R
\docType{package}
\name{paws.pi-package}
\alias{paws.pi}
\alias{paws.pi-package}
\title{paws.pi: AWS Performance Insights}
\description{
AWS Performance Insights enables you to monitor and
    explore different dimensions of database load based on data captured
    from a running RDS instance. The guide provides detailed information
    about Performance Insights data types, parameters and errors. For more
    information about Performance Insights capabilities see [Using Amazon
    RDS Performance
    Insights](http://docs.aws.amazon.com/AmazonRDS/latest/UserGuide/USER_PerfInsights.html)
    in the *Amazon RDS User Guide*.
}
\author{
\strong{Maintainer}: David Kretch \email{david.kretch@gmail.com}

Authors:
\itemize{
  \item Adam Banker \email{adam.banker39@gmail.com}
}

Other contributors:
\itemize{
  \item Amazon.com, Inc. [copyright holder]
}

}
\keyword{internal}
